Rename `config` package to `settings`
