+++ cli-api/src/main/kotlin/io/spine/protodata/cli/CommandLineInterface.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,11 +29,11 @@
 package io.spine.protodata.cli
 
 import com.google.protobuf.compiler.PluginProtos.CodeGeneratorRequest
-import io.spine.protodata.config.ConfigurationFormat.JSON
-import io.spine.protodata.config.ConfigurationFormat.PLAIN
-import io.spine.protodata.config.ConfigurationFormat.PROTO_JSON
-import io.spine.protodata.config.ConfigurationFormat.YAML
 import io.spine.protodata.plugin.Plugin
+import io.spine.protodata.settings.ConfigurationFormat.JSON
+import io.spine.protodata.settings.ConfigurationFormat.PLAIN
+import io.spine.protodata.settings.ConfigurationFormat.PROTO_JSON
+import io.spine.protodata.settings.ConfigurationFormat.YAML
 
 /**
  * The command-line parameter for specifying ProtoData plugins used in
@@ -120,6 +120,10 @@ public object UserClasspathParam : Parameter(
  * The command-line parameter for specifying the path to the file with custom
  * configuration for ProtoData.
  */
+@Deprecated(
+    message = "Use `ConfigDirParam` instead.",
+    replaceWith = ReplaceWith("ConfigDirParam")
+)
 public object ConfigFileParam : Parameter(
     name = "--configuration-file",
     shortName = "-c",
@@ -131,6 +135,27 @@ public object ConfigFileParam : Parameter(
         JSON files with Protobuf JSON format must have `.pb.json` extension.
         YAML files must have `.yml` or `.yaml` extension.
         Protobuf binary files must have `.pb` or `.bin` extension.
+        Messages must not be delimited.
+        """
+)
+
+/**
+ * The command-line parameter for specifying the path to the directory with
+ * setting files for ProtoData plugins.
+ */
+public object SettingsDirParam : Parameter (
+    name = "--settings-dir",
+    shortName = "-d",
+    help = """
+        A directory which contains setting files for ProtoData plugins.
+        
+        Setting files may be a JSON, a YAML, or a binary Protobuf file.
+        A name of the file must match the name of the plugin class, with the extension
+        corresponding to the format of the file:
+         * JSON files must have `.json` extension.
+         * JSON files with Protobuf JSON format must have `.pb.json` extension.
+         * YAML files must have `.yml` or `.yaml` extension.
+         * Protobuf binary files must have `.pb` or `.bin` extension.
         Messages must not be delimited.
         """
 )
