+++ compiler/src/test/kotlin/io/spine/protodata/backend/ConfigurationContextSpec.kt
@@ -29,13 +29,13 @@ package io.spine.protodata.backend
 import com.google.common.collect.ImmutableSet
 import io.kotest.matchers.collections.shouldContainExactly
 import io.spine.base.EventMessage
-import io.spine.protodata.config.ConfigurationFormat
-import io.spine.protodata.config.configFile
-import io.spine.protodata.config.event.FileConfigDiscovered
-import io.spine.protodata.config.event.RawConfigDiscovered
-import io.spine.protodata.config.event.fileConfigDiscovered
-import io.spine.protodata.config.event.rawConfigDiscovered
-import io.spine.protodata.config.rawConfig
+import io.spine.protodata.settings.ConfigurationFormat
+import io.spine.protodata.settings.configFile
+import io.spine.protodata.settings.event.FileConfigDiscovered
+import io.spine.protodata.settings.event.RawConfigDiscovered
+import io.spine.protodata.settings.event.fileConfigDiscovered
+import io.spine.protodata.settings.event.rawConfigDiscovered
+import io.spine.protodata.settings.rawConfig
 import io.spine.server.BoundedContext
 import io.spine.server.BoundedContextBuilder
 import io.spine.server.dispatch.DispatchOutcome
